test(preset-jekyll): use node test runner
